requirements.txt
coverage==4.2
coveralls==1.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.1
django-allauth==0.27.0
django-appconf==1.0.2
django-braces==1.9.0
django-cors-headers==1.1.0
django-coverage==1.2.4
django-imagekit==3.3
django-nose==1.4.4
django-oauth-toolkit==0.10.0
django-rest-auth==0.8.1
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
docopt==0.6.2
Factory==1.2
gunicorn==19.6.0
nose==1.3.7
oauthlib==1.0.3
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
pilkit==1.1.13
Pillow==3.3.1
psycopg2==2.6.2
ptyprocess==0.5.1
PyJWT==1.4.2
python-openid==2.2.5
python-resize-image==1.1.10
python-social-auth==0.2.21
requests==2.11.1
requests-oauthlib==0.6.2
static3==0.7.0
